// packages/client/src/generated/cosmwasm-codegen/Controller.client.ts
export interface IControllerAppQueryClient {
  moduleId: string;
  accountQueryClient: AbstractAccountQueryClient;
  _moduleAddress: string | undefined;
  status: (params: CamelCasedProperties<Extract<QueryMsg, {
    status: unknown;
  }>["status"]>) => Promise<StatusResponse>;
  config: () => Promise<ConfigResponse>;
  userPosition: (params: CamelCasedProperties<Extract<QueryMsg, {
    user_position: unknown;
  }>["user_position"]>) => Promise<UserPositionResponse>;
  participants: () => Promise<ParticipantsResponse>;
  connectSigningClient: (signingClient: SigningCosmWasmClient, address: string) => ControllerAppClient;
  getAddress: () => Promise<string>;
}
export class ControllerAppQueryClient implements IControllerAppQueryClient {
  accountQueryClient: AbstractAccountQueryClient;
  moduleId: string;
  _moduleAddress: string | undefined;

  constructor({
    abstractQueryClient,
    accountId,
    managerAddress,
    proxyAddress,
    moduleId
  }: {
    abstractQueryClient: AbstractQueryClient;
    accountId: AbstractAccountId;
    managerAddress: string;
    proxyAddress: string;
    moduleId: string;
  }) {
    this.accountQueryClient = new AbstractAccountQueryClient({
      abstract: abstractQueryClient,
      accountId,
      managerAddress,
      proxyAddress
    });
    this.moduleId = moduleId;
    this.status = this.status.bind(this);
    this.config = this.config.bind(this);
    this.userPosition = this.userPosition.bind(this);
    this.participants = this.participants.bind(this);
  }

  status = async (params: CamelCasedProperties<Extract<QueryMsg, {
    status: unknown;
  }>["status"]>): Promise<StatusResponse> => {
    return this._query(ControllerQueryMsgBuilder.status(params));
  };
  config = async (): Promise<ConfigResponse> => {
    return this._query(ControllerQueryMsgBuilder.config());
  };
  userPosition = async (params: CamelCasedProperties<Extract<QueryMsg, {
    user_position: unknown;
  }>["user_position"]>): Promise<UserPositionResponse> => {
    return this._query(ControllerQueryMsgBuilder.userPosition(params));
  };
  participants = async (): Promise<ParticipantsResponse> => {
    return this._query(ControllerQueryMsgBuilder.participants());
  };
  getAddress = async (): Promise<string> => {
    if (!this._moduleAddress) {
      const address = await this.accountQueryClient.getModuleAddress(this.moduleId);

      if (address === null) {
        throw new Error(`Module ${this.moduleId} not installed`);
      }

      this._moduleAddress = address;
    }

    return this._moduleAddress!;
  };
  connectSigningClient = (signingClient: SigningCosmWasmClient, address: string): ControllerAppClient => {
    return new ControllerAppClient({
      accountId: this.accountQueryClient.accountId,
      managerAddress: this.accountQueryClient.managerAddress,
      proxyAddress: this.accountQueryClient.proxyAddress,
      moduleId: this.moduleId,
      abstractClient: this.accountQueryClient.abstract.connectSigningClient(signingClient, address)
    });
  };
  _query = async (queryMsg: QueryMsg): Promise<any> => {
    return this.accountQueryClient.queryModule({
      moduleId: this.moduleId,
      moduleType: "adapter",
      queryMsg
    });
  };
}
export interface IControllerAppClient extends IControllerAppQueryClient {
  accountClient: AbstractAccountClient;
  setStatus: (params: CamelCasedProperties<Extract<ExecuteMsg, {
    set_status: unknown;
  }>["set_status"]>, fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
  updateConfig: (fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
  join: (fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
  rollDice: (fee?: number | StdFee | "auto", memo?: string, _funds?: Coin[]) => Promise<ExecuteResult>;
}
export class ControllerAppClient extends ControllerAppQueryClient implements IControllerAppClient {
  accountClient: AbstractAccountClient;

  constructor({
    abstractClient,
    accountId,
    managerAddress,
    proxyAddress,
    moduleId
  }: {
    abstractClient: AbstractClient;
    accountId: AbstractAccountId;
    managerAddress: string;
    proxyAddress: string;
    moduleId: string;
  }) {
    super({
      abstractQueryClient: abstractClient,
      accountId,
      managerAddress,
      proxyAddress,
      moduleId
    });
    this.accountClient = AbstractAccountClient.fromQueryClient(this.accountQueryClient, abstractClient);
    this.setStatus = this.setStatus.bind(this);
    this.updateConfig = this.updateConfig.bind(this);
    this.join = this.join.bind(this);
    this.rollDice = this.rollDice.bind(this);
  }

  setStatus = async (params: CamelCasedProperties<Extract<ExecuteMsg, {
    set_status: unknown;
  }>["set_status"]>, fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(ControllerExecuteMsgBuilder.setStatus(params), fee, memo, _funds);
  };
  updateConfig = async (fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(ControllerExecuteMsgBuilder.updateConfig(), fee, memo, _funds);
  };
  join = async (fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(ControllerExecuteMsgBuilder.join(), fee, memo, _funds);
  };
  rollDice = async (fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    return this._execute(ControllerExecuteMsgBuilder.rollDice(), fee, memo, _funds);
  };
  _execute = async (msg: ExecuteMsg, fee: number | StdFee | "auto" = "auto", memo?: string, _funds?: Coin[]): Promise<ExecuteResult> => {
    const moduleMsg: AdapterExecuteMsg<ExecuteMsg> = AdapterExecuteMsgFactory.executeAdapter({
      request: msg,
      proxyAddress: this.accountQueryClient.proxyAddress
    });
    return await this.accountClient.abstract.client.execute(this.accountClient.sender, await this.getAddress(), moduleMsg, fee, memo, _funds);
  };
}